删除角色时判断role——user表中是否有记录

+++ src/src/main/java/zy/news/web/service/impl/SvrImpAuthRole.java
@@ -38,7 +38,10 @@ public class SvrImpAuthRole implements IAuthRole {
         List<Long> ids = new ArrayList<>(1);
         ids.add(id);
         if (mapper.countByExampleRolePerms(ids) > 0) {
-            throw new WarningException("角色-权限中仍存在待删除角色信息，请先取消角色关联绑定，再选择此角色进行删除。");
+            throw new WarningException("角色-权限中仍存在待删除角色信息，请先取消权限角色关联绑定，再选择此角色进行删除。");
+        }
+        if (mapper.countRoleUsers(ids) > 0) {
+            throw new WarningException("角色-用户中仍存在待删除角色信息，请先取消用户角色关联绑定，再选择此角色进行删除。");
         }
         int count = mapper.deleteByPrimaryKey(id);
         return count;
